PEP8 fixes; cleaned up the example cfg file

+++ pb_tool/pb_tool.py
@@ -2,7 +2,6 @@ __author__ = 'gsherman'
 import os
 import sys
 import subprocess
-import zipfile
 import shutil
 import errno
 import ConfigParser
@@ -10,11 +9,20 @@ import ConfigParser
 
 import click
 
+
 @click.group()
 def cli():
     """Simple Python tool to compile and deploy a QGIS plugin.
     For help on a command use --help after the command:
-    pb_tool deploy --help"""
+    pb_tool deploy --help.
+
+    pb_tool requires a configuration file (pb_tool.cfg) that
+    declares the files and resources used in your plugin. Plugin Builder
+    2.6.0 creates a config file when you generate a new plugin template.
+
+    See
+    https://github.com/g-sherman/plugin_build_tool/blob/master/test_plugin/pb_tool.cfg
+    for an example."""
     pass
 
 
@@ -29,10 +37,12 @@ def get_install_files():
     #click.echo(install_files)
     return install_files
 
+
 @cli.command()
 def version():
     """Return the version of pb_tool and exit"""
-    click.echo("1.0 2014-09-25")
+    click.echo("1.0a1, 2014-10-01")
+
 
 @cli.command()
 def deploy():
@@ -45,36 +55,58 @@ def deploy():
         * Remove your currently deployed version
         * Compile the ui and resource files
         * Build the help docs
-        * Copy everything to the .qgis/python/plugins directory
+        * Copy everything to your .qgis2/python/plugins directory
         """
 
-        plugin_dir = os.path.join(get_plugin_directory(), config().get('plugin', 'name'))
-        print ">> Deploying to {} <<".format(plugin_dir)
-        # clean the deployment
-        clean_deployment(False)
-        # compile to make sure everything is fresh
-        print 'Compiling to make sure install is clean'
-        compile_files()
-        build_docs()
-        install_files = get_install_files()
-        # make the plugin directory if it doesn't exist
-        if not os.path.exists(plugin_dir):
-            os.mkdir(plugin_dir)
+        if click.confirm("Proceed?"):
 
-        try:
-            for file in install_files:
-                print "Copying {}".format(file)
-                shutil.copy(file, os.path.join(plugin_dir, file))
-            help_src = config().get('help', 'dir')
-            help_target = os.path.join(
-                get_plugin_directory(),
-                config().get('plugin', 'name'),
-                config().get('help', 'target'))
-            print "Copying {} to {}".format(help_src, help_target)
-            shutil.copytree(help_src, help_target)
-        except OSError as oops:
-            print "Error copying files: {}, {}".format(file, oops.strerror)
-            print "To ensure proper deployment, delete the deployed plugin before continuing."
+            plugin_dir = os.path.join(get_plugin_directory(), config().get('plugin', 'name'))
+            # clean the deployment
+            clean_deployment(False)
+            print "Deploying to {}".format(plugin_dir)
+            # compile to make sure everything is fresh
+            print 'Compiling to make sure install is clean'
+            compile_files()
+            build_docs()
+            install_files = get_install_files()
+            # make the plugin directory if it doesn't exist
+            if not os.path.exists(plugin_dir):
+                os.mkdir(plugin_dir)
+
+            fail = False
+            try:
+                for file in install_files:
+                    print "Copying {}".format(file)
+                    shutil.copy(file, os.path.join(plugin_dir, file))
+                # copy extra dirs
+            except OSError as oops:
+                print "Error copying files: {}, {}".format(file, oops.strerror)
+                fail = True
+            try:
+                extra_dirs = config().get('files', 'extra_dirs').split()
+                #print "EXTRA DIRS: {}".format(extra_dirs)
+                for xdir in extra_dirs:
+                    print "Copying contents of {} to {}".format(xdir, plugin_dir)
+                    shutil.copytree(xdir, "{}/{}".format(
+                        plugin_dir, xdir))
+            except OSError as oops:
+                print "Error copying directory: {}, {}".format(xdir, oops.strerror)
+                fail = True
+            try:
+                help_src = config().get('help', 'dir')
+                help_target = os.path.join(
+                    get_plugin_directory(),
+                    config().get('plugin', 'name'),
+                    config().get('help', 'target'))
+                print "Copying {} to {}".format(help_src, help_target)
+                shutil.copytree(help_src, help_target)
+            except OSError as oops:
+                print "Error copying help files: {}, {}".format(help_src, oops.strerror)
+                fail = True
+            if fail:
+                print "\nOne or more files/directories failed to deploy."
+                print "To ensure proper deployment, try using dclean to delete"
+                print "the plugin before deploying."
 
 
 def clean_deployment(ask_first=True):
@@ -88,7 +120,7 @@ def clean_deployment(ask_first=True):
         proceed = True
 
     if proceed:
-        click.echo('Removing from {}'.format(plugin_dir))
+        click.echo('Removing plugin from {}'.format(plugin_dir))
         try:
             shutil.rmtree(plugin_dir)
             return True
@@ -97,6 +129,7 @@ def clean_deployment(ask_first=True):
     else:
         click.echo('Plugin was not deleted')
     return False
+
 
 @cli.command()
 def clean_docs():
@@ -116,11 +149,13 @@ def clean_docs():
     else:
         print "No help directory exists in the current directory"
 
+
 @cli.command()
 def dclean():
     """ Remove the deployed plugin from the .qgis2/python/plugins directory
     """
     clean_deployment()
+
 
 @cli.command()
 def clean():
@@ -149,6 +184,7 @@ def doc():
     """ Build HTML version of the help files using sphinx"""
     build_docs()
 
+
 def build_docs():
     """ Build the docs using sphinx"""
     if os.path.exists('help'):
@@ -163,6 +199,7 @@ def build_docs():
         os.chdir(cwd)
     else:
         print "No help directory exists in the current directory"
+
 
 @cli.command()
 def zip():
@@ -189,11 +226,57 @@ def zip():
             click.echo("Your config file is missing the plugin name (name=parameter)")
 
 
+@cli.command()
 def validate():
     """
-    Validate existence of needed commands
+    Check the pb_tool.cfg file for mandatory sections/files
     """
-    pass
+    valid = True
+    cfg = config()
+    if not check_cfg(cfg, 'plugin', 'name'):
+        valid = False
+    if not check_cfg(cfg, 'files', 'python_files'):
+        valid = False
+    if not check_cfg(cfg, 'files', 'main_dialog'):
+        valid = False
+    if not check_cfg(cfg, 'files', 'resource_files'):
+        valid = False
+    if not check_cfg(cfg, 'files', 'extras'):
+        valid = False
+    if not check_cfg(cfg, 'help', 'dir'):
+        valid = False
+    if not check_cfg(cfg, 'help', 'target'):
+        valid = False
+
+    if valid:
+        print "Your pb_tool.cfg file is valid and contains all mandatory items"
+    else:
+        print "Your pb_tool.cfg file is invalid"
+
+
+@cli.command()
+def list():
+    """ List the contents of the configuration file """
+    if os.path.exists('pb_tool.cfg'):
+        with open('pb_tool.cfg') as config:
+            for line in config:
+                print line[:-1]
+    else:
+        print "There is no pb_tool.cfg file in the current directory"
+        print "We can't do anything without it"
+
+
+def check_cfg(cfg, section, name):
+    try:
+        cfg.get(section, name)
+        return True
+    except ConfigParser.NoOptionError as oops:
+        print oops.message
+    except ConfigParser.NoSectionError as secoops:
+        print "Missing section '{}' when looking for option '{}'".format(
+              section, name)
+    return False
+
 
 def config():
     """
@@ -209,6 +292,7 @@ def config():
         print "We can't do anything without it"
         sys.exit(1)
 
+
 def compiled_ui():
     cfg = config()
     try:
@@ -223,6 +307,7 @@ def compiled_ui():
         print oops.message
         sys.exit(1)
 
+
 def compiled_resource():
     cfg = config()
     try:
@@ -236,6 +321,7 @@ def compiled_resource():
     except ConfigParser.NoSectionError as oops:
         print oops.message
         sys.exit(1)
+
 
 def compile_files():
     # Compile all ui and resource files
@@ -289,6 +375,7 @@ def copy(source, destination):
             shutil.copy(source, destination)
         else:
             print('Directory not copied. Error: %s' % e)
+
 
 def get_plugin_directory():
     if sys.platform == 'win32':
